+++ src/lib/utils/index.ts
@@ -19,7 +19,7 @@ export {
 } from "./fs";
 export type { IfInternal, NeverIfInternal, Chars } from "./general";
 export { assertNever } from "./general";
-export { CallbackLogger, ConsoleLogger, Logger, LogLevel } from "./loggers";
+export { ConsoleLogger, Logger, LogLevel } from "./loggers";
 export { DefaultMap } from "./map";
 export {
     ArgumentsReader,